requirements.txt
aiohttp==3.8.5
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
av==8.0.3
beautifulsoup4==4.11.2
boto3==1.26.36
botocore==1.29.36
charset-normalizer==2.1.1
click==8.1.3
decord==0.6.0
docker-pycreds==0.4.0
dominate==2.6.0
ego4d==1.2.5
einops==0.4.1
ffmpeg==1.4
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
gdown==4.6.3
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.11.0
humanize==3.12.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.5
opencv-python==4.6.0.66
packaging==21.0
pandas==1.1.5
pathtools==0.1.2
Pillow==9.3.0
prettytable==3.4.0
protobuf==3.18.3
psutil==5.8.0
pymongo==4.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.10.31
requests==2.31.0
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.10.0
sentry-sdk==1.14.0
setproctitle==1.3.2
Shapely==1.8.4
six==1.16.0
smmap==5.0.0
soupsieve==2.4
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
threadpoolctl==3.1.0
timm==0.4.12
tokenizers==0.13.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
transformers==4.30.0
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.9
wcwidth==0.2.5
Werkzeug==2.2.3
yacs==0.1.8
yarl==1.8.1
